Refactor: Add theme selection to user profile dropdown
